Fix warnings
Former-commit-id: 252ec410d6e4bf47d9ab669594eecb91da593d22
Former-commit-id: a19a9288ae1d89d5604cfaba2e035f2eaea424a9

+++ plugins/sdl/src/minko/Canvas.cpp
@@ -221,25 +221,25 @@ Canvas::initContext(SDL_Window* window, unsigned int width, unsigned int height)
 uint
 Canvas::x()
 {
-	return _data->get<math::Vector4::Ptr>("canvas.viewport")->x();
+	return (uint)_data->get<math::Vector4::Ptr>("canvas.viewport")->x();
 }
 
 uint
 Canvas::y()
 {
-	return _data->get<math::Vector4::Ptr>("canvas.viewport")->y();
+	return (uint)_data->get<math::Vector4::Ptr>("canvas.viewport")->y();
 }
 
 uint
 Canvas::width()
 {
-	return _data->get<math::Vector4::Ptr>("canvas.viewport")->z();
+	return (uint)_data->get<math::Vector4::Ptr>("canvas.viewport")->z();
 }
 
 uint
 Canvas::height()
 {
-	return _data->get<math::Vector4::Ptr>("canvas.viewport")->w();
+	return (uint)_data->get<math::Vector4::Ptr>("canvas.viewport")->w();
 }
 
 void
